requirements.txt
alembic==1.18.4
celery==5.5.3
Django==5.2.7
easysnmp==0.2.6
fastapi==0.135.1
passlib[bcrypt]==1.7.4
pydantic==2.12.5
pydantic-settings==2.11.0
PyMySQL==1.1.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.22
redis==6.4.0
routeros-api==0.21.0
SQLAlchemy==2.0.48
uvicorn==0.41.0
